Code cleanup, refactor, updated logging.
Deleted various unused import statements, updated logging code, added
environment variables for optional namespace deletion/cleanup.
Added option to deploy spark pods without the need for oshinko.
Jgrafzahl tests now work under any namespace.

Maven setup for aop logs using jcabi aspects implemented

Added aspect logging to all modules

Fixed deployment bug in s3source

Added better exception handling with meaningful error messages

Removed unused maven dependencies

+++ test-s3source/src/test/java/com/redhat/xpaas/Setup.java
@@ -1,43 +1,37 @@
 package com.redhat.xpaas;
 
+import com.redhat.xpaas.logger.Loggable;
 import com.redhat.xpaas.openshift.OpenshiftUtil;
 import com.redhat.xpaas.rad.S3Source.api.S3SourceWebUI;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+
+import java.util.concurrent.TimeoutException;
 
 import static com.redhat.xpaas.rad.S3Source.deployment.S3Source.deployS3Source;
 
+@Loggable(project = "S3Source")
 public class Setup {
-  private Logger log = LoggerFactory.getLogger(Setup.class);
-  private String NAMESPACE = RadConfiguration.masterNamespace();
   private static final OpenshiftUtil openshift = OpenshiftUtil.getInstance();
+  private String NAMESPACE = RadConfiguration.masterNamespace();
   private static S3SourceWebUI S3Source;
 
-  S3SourceWebUI initializeApplications() {
-    Logger log = LoggerFactory.getLogger(WebUITest.class);
-    log.info("project=s3source action=creating-new-namespace status=START");
+  public S3SourceWebUI initializeApplications() throws TimeoutException, InterruptedException {
     initializeProject();
-    log.info("project=s3source action=creating-new-namespace status=FINISH");
-    log.info("project=s3source action=deploy-s3source status=START");
     S3Source = deployS3Source();
-    log.info("project=s3source action=deploy-s3source status=FINISH");
     return S3Source;
   }
 
-  void cleanUp() {
+  public void cleanUp() {
     if(S3Source != null){
-      log.info("project=s3source action=shutting-down-webdrivers status=START");
       S3Source.webDriverCleanup();
-      log.info("project=s3source action=shutting-down-webdrivers status=FINISH");
     }
-    log.info("project=s3source action=deleting-namespace status=START");
-    openshift.deleteProject(NAMESPACE);
-    log.info("project=s3source action=shutting-down-webdrivers status=FINISH");
+
+    if(RadConfiguration.deleteNamespaceAfterTests()){
+      openshift.deleteProject(NAMESPACE);
+    }
   }
 
   private void initializeProject(){
-    OpenshiftUtil.getInstance().createProject(NAMESPACE, true);
+    OpenshiftUtil.getInstance().createProject(NAMESPACE, RadConfiguration.recreateNamespace());
   }
-
 
 }
